MNT: move enum options out to a separate module
